+++ src/NukiConstants.h
@@ -36,16 +36,6 @@ enum class BatteryType : uint8_t {
   Lithium           = 0x02
 };
 
-enum class LoggingType : uint8_t {
-  LoggingEnabled            = 0x01,
-  LockAction                = 0x02,
-  Calibration               = 0x03,
-  InitializationRun         = 0x04,
-  KeypadAction              = 0x05,
-  DoorSensor                = 0x06,
-  DoorSensorLoggingEnabled  = 0x07
-};
-
 enum class AdvertisingMode : uint8_t {
   Automatic                 = 0x00,
   Normal                    = 0x01,
@@ -122,6 +112,13 @@ enum class Command : uint16_t {
   KeypadAction	                = 0x0048,
   ContinuousModeAction     	     = 0x0057, // Opener only
   SimpleLockAction	            = 0x0100
+};
+
+enum class AuthorizationIdType : uint8_t {
+  App = 0,
+  Bridge = 1,
+  Fob = 2,
+  Keypad = 3
 };
 
 enum class TimeZoneId : uint16_t {
@@ -266,21 +263,6 @@ struct __attribute__((packed)) UpdatedKeypadEntry {
   uint8_t allowedFromTimeMin;
   uint8_t allowedUntilTimeHour;
   uint8_t allowedUntilTimeMin;
-};
-
-
-struct __attribute__((packed)) LogEntry {
-  uint32_t index;
-  uint16_t timeStampYear;
-  uint8_t timeStampMonth;
-  uint8_t timeStampDay;
-  uint8_t timeStampHour;
-  uint8_t timeStampMinute;
-  uint8_t timeStampSecond;
-  uint32_t authId;
-  uint8_t name[32];
-  Nuki::LoggingType loggingType;
-  uint8_t data[5];
 };
 
 struct __attribute__((packed)) AuthorizationEntry {