+++ src/NukiOpener.cpp
@@ -1,5 +1,6 @@
 #include "NukiOpener.h"
 #include "NukiUtils.h"
+#include "NukiOpenerUtils.h"
 
 namespace NukiOpener {
 NukiOpener::NukiOpener(const std::string& deviceName, const uint32_t deviceId)
@@ -39,6 +40,7 @@ Nuki::CmdResult NukiOpener::requestOpenerState(OpenerState* state) {
   Action action;
   uint16_t payload = (uint16_t)Command::KeyturnerStates;
 
+  memset(&action, 0, sizeof(action));
   action.cmdType = Nuki::CommandType::Command;
   action.command = Command::RequestData;
   memcpy(&action.payload[0], &payload, sizeof(payload));
@@ -61,6 +63,7 @@ Nuki::CmdResult NukiOpener::requestBatteryReport(BatteryReport* retrievedBattery
   Action action;
   uint16_t payload = (uint16_t)Command::BatteryReport;
 
+  memset(&action, 0, sizeof(action));
   action.cmdType = Nuki::CommandType::Command;
   action.command = Command::RequestData;
   memcpy(&action.payload[0], &payload, sizeof(payload));
@@ -77,6 +80,7 @@ Nuki::CmdResult NukiOpener::requestBatteryReport(BatteryReport* retrievedBattery
 Nuki::CmdResult NukiOpener::requestConfig(Config* retrievedConfig) {
   Action action;
 
+  memset(&action, 0, sizeof(action));
   action.cmdType = Nuki::CommandType::CommandWithChallenge;
   action.command = Command::RequestConfig;
 
@@ -90,6 +94,7 @@ Nuki::CmdResult NukiOpener::requestConfig(Config* retrievedConfig) {
 Nuki::CmdResult NukiOpener::requestAdvancedConfig(AdvancedConfig* retrievedAdvancedConfig) {
   Action action;
 
+  memset(&action, 0, sizeof(action));
   action.cmdType = Nuki::CommandType::CommandWithChallenge;
   action.command = Command::RequestAdvancedConfig;
 
@@ -223,8 +228,7 @@ CmdResult NukiOpener::enableLedFlash(const bool enable) {
   return result;
 }
 
-CmdResult NukiOpener::setSoundLevel(const uint8_t value)
-{
+CmdResult NukiOpener::setSoundLevel(const uint8_t value) {
   AdvancedConfig oldConfig;
   Nuki::CmdResult result = requestAdvancedConfig(&oldConfig);
   if (result == Nuki::CmdResult::Success) {
@@ -324,6 +328,32 @@ void NukiOpener::getTimeControlEntries(std::list<TimeControlEntry>* requestedTim
   }
 }
 
+void NukiOpener::getLogEntries(std::list<LogEntry>* requestedLogEntries) {
+  requestedLogEntries->clear();
+
+  for (const auto& it : listOfLogEntries) {
+    requestedLogEntries->push_back(it);
+  }
+}
+
+Nuki::CmdResult NukiOpener::retrieveLogEntries(const uint32_t startIndex, const uint16_t count, const uint8_t sortOrder, bool const totalCount) {
+  Action action;
+  unsigned char payload[8] = {0};
+  memcpy(payload, &startIndex, 4);
+  memcpy(&payload[4], &count, 2);
+  memcpy(&payload[6], &sortOrder, 1);
+  memcpy(&payload[7], &totalCount, 1);
+
+  action.cmdType = Nuki::CommandType::CommandWithChallengeAndPin;
+  action.command = Command::RequestLogEntries;
+  memcpy(action.payload, &payload, sizeof(payload));
+  action.payloadLen = sizeof(payload);
+
+  listOfLogEntries.clear();
+
+  return executeAction(action);
+}
+
 bool NukiOpener::isBatteryCritical() {
   return openerState.criticalBatteryState & 1;
 }
@@ -420,7 +450,7 @@ void NukiOpener::handleReturnMessage(Command returnCode, unsigned char* data, ui
       printBuffer((byte*)data, dataLen, false, "keyturnerStates");
       memcpy(&openerState, data, sizeof(openerState));
       #ifdef DEBUG_NUKI_READABLE_DATA
-      logKeyturnerState(keyTurnerState);
+      logKeyturnerState(openerState);
       #endif
       break;
     }
@@ -455,6 +485,16 @@ void NukiOpener::handleReturnMessage(Command returnCode, unsigned char* data, ui
       listOfTimeControlEntries.push_back(timeControlEntry);
       break;
     }
+    case Command::LogEntry : {
+      printBuffer((byte*)data, dataLen, false, "logEntry");
+      LogEntry logEntry;
+      memcpy(&logEntry, data, sizeof(logEntry));
+      listOfLogEntries.push_back(logEntry);
+      #ifdef DEBUG_NUKI_READABLE_DATA
+      logLogEntry(logEntry);
+      #endif
+      break;
+    }
     default:
       NukiBle::handleReturnMessage(returnCode, data, dataLen);
   }
@@ -462,7 +502,7 @@ void NukiOpener::handleReturnMessage(Command returnCode, unsigned char* data, ui
 }
 
 void NukiOpener::logErrorCode(uint8_t errorCode) {
-  NukiOpener::logErrorCode(errorCode);
+  logOpenerErrorCode(errorCode);
 }