+++ src/NukiOpenerConstants.h
@@ -98,6 +98,7 @@ enum class State : uint8_t {
   Uninitialized   = 0x00,
   PairingMode     = 0x01,
   DoorMode        = 0x02,
+  ContinuousMode  = 0x03,
   MaintenanceMode = 0x04
 };
 
@@ -175,6 +176,9 @@ struct __attribute__((packed)) Config {
   unsigned char firmwareVersion[3];
   unsigned char hardwareRevision[2];
   TimeZoneId timeZoneId;
+  uint8_t undocumented1;
+  uint8_t undocumented2;
+  uint8_t hasKeypadV2;
 };
 
 struct __attribute__((packed)) NewConfig {
@@ -272,6 +276,60 @@ struct __attribute__((packed)) NewTimeControlEntry {
   LockAction lockAction;
 };
 
+enum class LoggingType : uint8_t {
+  LoggingEnabled            = 0x01,
+  LockAction                = 0x02,
+  Calibration               = 0x03,
+  KeypadAction              = 0x05,
+  DoorbellRecognition       = 0x06
+};
+
+struct __attribute__((packed)) LogEntry {
+  uint32_t index;
+  uint16_t timeStampYear;
+  uint8_t timeStampMonth;
+  uint8_t timeStampDay;
+  uint8_t timeStampHour;
+  uint8_t timeStampMinute;
+  uint8_t timeStampSecond;
+  uint32_t authId;
+  uint8_t name[32];
+  LoggingType loggingType;
+  uint8_t data[8];
+};
+
+inline void lockactionToString(const LockAction action, char* str) {
+  switch (action) {
+    case LockAction::ActivateRTO:
+      strcpy(str, "ActivateRTO");
+      break;
+    case LockAction::DeactivateRTO:
+      strcpy(str, "DeactivateRTO");
+      break;
+    case LockAction::ElectricStrikeActuation:
+      strcpy(str, "ElectricStrikeActuation");
+      break;
+    case LockAction::ActivateCM:
+      strcpy(str, "ActivateCM");
+      break;
+    case LockAction::DeactivateCM:
+      strcpy(str, "DeactivateCM");
+      break;
+    case LockAction::FobAction1:
+      strcpy(str, "FobAction1");
+      break;
+    case LockAction::FobAction2:
+      strcpy(str, "FobAction2");
+      break;
+    case LockAction::FobAction3:
+      strcpy(str, "FobAction3");
+      break;
+    default:
+      strcpy(str, "Unknown");
+      break;
+  }
+}
+
 inline void lockstateToString(const LockState state, char* str) {
   switch (state) {
     case LockState::Uncalibrated:
@@ -368,6 +426,29 @@ inline void doorSensorStateToString(const DoorSensorState state, char* str) {
       break;
     default:
       strcpy(str, "undefined");
+      break;
+  }
+}
+
+inline void loggingTypeToString(const LoggingType state, char* str) {
+  switch (state) {
+    case LoggingType::LoggingEnabled:
+      strcpy(str, "LoggingEnabled");
+      break;
+    case LoggingType::LockAction:
+      strcpy(str, "LockAction");
+      break;
+    case LoggingType::Calibration:
+      strcpy(str, "Calibration");
+      break;
+    case LoggingType::KeypadAction:
+      strcpy(str, "KeypadAction");
+      break;
+    case LoggingType::DoorbellRecognition:
+      strcpy(str, "DoorbellRecognition");
+      break;
+    default:
+      strcpy(str, "Unknown");
       break;
   }
 }