+++ src/NukiUtils.h
@@ -24,6 +24,7 @@ namespace Nuki {
 
 void printBuffer(const byte* buff, const uint8_t size, const boolean asChars, const char* header);
 bool isCharArrayNotEmpty(unsigned char* array, uint16_t len);
+bool isCharArrayEmpty(unsigned char* array, uint16_t len);
 bool compareCharArray(unsigned char* a, unsigned char* b, uint8_t len);
 int encode(unsigned char* output, unsigned char* input, unsigned long long len, unsigned char* nonce, unsigned char* keyS);
 int decode(unsigned char* output, unsigned char* input, unsigned long long len, unsigned char* nonce, unsigned char* keyS);